update: shopping item 調整した

+++ server/app/controllers/shopping_items_controller.rb
@@ -1,8 +1,10 @@
 class ShoppingItemsController < ApplicationController
   include ActionController::MimeResponds
+  before_action :set_shopping_item, only: [ :update, :destroy ]
+  before_action :check_different_user, only: [ :update, :destroy ]
 
   def show
-    @shopping_item = ShoppingItem.find(params[:id])
+    set_shopping_item
 
     respond_to do |format|
       format.json { render json: @shopping_item }
@@ -17,31 +19,42 @@ class ShoppingItemsController < ApplicationController
     end
   end
 
-  def update
-    @shopping_item = ShoppingItem.find(params[:id])
-    if @shopping_item.update(shopping_item_params)
-      head :ok
-    else
-      render json: { error: 'Failed to update the item' }, status: :unprocessable_entity
-    end
-  end
-
   def create
     @shopping_item = ShoppingItem.new(shopping_item_params)
+    @shopping_item.user_id = current_user.id
+
     if @shopping_item.save
       head :ok
     else
-      render json: { error: 'Failed to create the item' }, status: :unprocessable_entity
+      render json: { error: "Failed to create the item. Reason: #{@shopping_item.errors.full_messages}" }, status: :unprocessable_entity
+    end
+  end
+
+  def update
+    set_shopping_item
+    check_different_user
+
+    if @shopping_item.update(shopping_item_params)
+      head :ok
+    else
+      render json: { error: "Failed to update the item. Reason: #{@shopping_item.errors.full_messages}" }, status: :unprocessable_entity
     end
   end
 
   def destroy
-    @shopping_item = ShoppingItem.find(params[:id])
+    set_shopping_item
+    check_different_user
+
+    if @shopping_item.user_id != current_user.id
+      render json: { error: "Cannot Delete Different User Shopping Item" }, status: :unprocessable_entity
+      return
+    end
+
     if @shopping_item.destroy
       head :ok
     else
-      render json: { error: 'Failed to delete the item' }, status: :unprocessable_entity
-    end      
+      render json: { error: "Failed to delete the item. Reason: #{@shopping_item.errors.full_messages}" }, status: :unprocessable_entity
+    end
   end
 
   private
@@ -49,4 +62,11 @@ class ShoppingItemsController < ApplicationController
       params.require(:shopping_item).permit(:item, :checked)
     end
 
+    def set_shopping_item
+      @shopping_item = ShoppingItem.find(params[:id])
+    end
+
+    def check_different_user
+      render json: { error: "Cannot Act Different User Shopping Item" }, status: :unprocessable_entity unless @shopping_item.user_id == current_user.id
+    end
 end
